+++ jdbc/src/main/java/org/springframework/jdbc/core/JdbcTemplate.java
@@ -8,7 +8,7 @@ import java.util.ArrayList;
 import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.SqlQueryException;
+import org.springframework.jdbc.SqlQueryException;
 
 public class JdbcTemplate {
